Pipeline structuring and inspection breaking off from analysis
